requirements.txt
certifi==2019.6.16
cycler==0.10.0
decorator==4.4.0
imageio==2.5.0
kiwisolver==1.1.0
matplotlib==3.1.1
networkx==2.3
numpy==1.17.2
opencv-python==4.1.1.26
Pillow==6.1.0
pyparsing==2.4.2
python-dateutil==2.8.0
PyWavelets==1.0.3
PyYAML==5.1.2
scikit-image==0.15.0
scipy==1.3.1
six==1.12.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.35.0
